Force an app change before mapping a route
Route changes result in the CC sending a `diego.desire.app` message to
☆NSYNC. ☆NSYNC converts that message to a DesiredLRPCreateRequest and
forwards it on to the receptor who forwards it to the BBS...

The current BBS code will implicitly convert a 'create' into an 'update'
if only the instance count, routes, or annotation changes and will
reject any other updates.

When an app change exists prior to changing routes, the desire app
message from the CC can result in a desired LRP that has the same
process guid but different actions. When that happens, the request is
rejected.

Signed-off-by: Michael Fraenkel <fraenkel@us.ibm.com>